refactor: share mappers
