+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/main/java/io/gravitee/rest/api/management/v2/rest/mapper/ApiMapper.java
@@ -33,7 +33,7 @@ import org.mapstruct.Mapping;
 import org.mapstruct.Named;
 import org.mapstruct.factory.Mappers;
 
-@Mapper
+@Mapper(uses = { DateMapper.class, FlowMapper.class })
 public interface ApiMapper {
     ApiMapper INSTANCE = Mappers.getMapper(ApiMapper.class);
 
@@ -60,13 +60,6 @@ public interface ApiMapper {
     NewApiEntity convert(CreateApiV4 api);
 
     io.gravitee.rest.api.management.v2.rest.model.ApiV2 map(io.gravitee.rest.api.model.api.ApiEntity apiEntity);
-
-    // Flow
-    @Mapping(target = "selectors", qualifiedByName = "fromSelectors")
-    FlowV4 map(io.gravitee.definition.model.v4.flow.Flow flow);
-
-    @Mapping(target = "selectors", qualifiedByName = "toSelectors")
-    io.gravitee.definition.model.v4.flow.Flow map(FlowV4 flow);
 
     // DefinitionVersion
     io.gravitee.definition.model.DefinitionVersion map(DefinitionVersion definitionVersion);
@@ -154,78 +147,5 @@ public interface ApiMapper {
                 return null;
             })
             .collect(Collectors.toList());
-    }
-
-    // Selectors
-    HttpSelector map(io.gravitee.definition.model.v4.flow.selector.HttpSelector selector);
-    ConditionSelector map(io.gravitee.definition.model.v4.flow.selector.ConditionSelector selector);
-    ChannelSelector map(io.gravitee.definition.model.v4.flow.selector.ChannelSelector selector);
-
-    io.gravitee.definition.model.v4.flow.selector.HttpSelector map(HttpSelector selector);
-    io.gravitee.definition.model.v4.flow.selector.ConditionSelector map(ConditionSelector selector);
-    io.gravitee.definition.model.v4.flow.selector.ChannelSelector map(ChannelSelector selector);
-
-    @Named("toSelectors")
-    default List<io.gravitee.definition.model.v4.flow.selector.Selector> toSelectors(List<Selector> selectors) {
-        if (Objects.isNull(selectors)) {
-            return new ArrayList<>();
-        }
-        return selectors
-            .stream()
-            .map(selector -> {
-                if (selector == null) {
-                    return null;
-                }
-                if (selector.getActualInstance() instanceof io.gravitee.rest.api.management.v2.rest.model.HttpSelector) {
-                    return this.map(selector.getHttpSelector());
-                }
-                if (selector.getActualInstance() instanceof io.gravitee.rest.api.management.v2.rest.model.ConditionSelector) {
-                    return this.map(selector.getConditionSelector());
-                }
-                if (selector.getActualInstance() instanceof io.gravitee.rest.api.management.v2.rest.model.ChannelSelector) {
-                    return this.map(selector.getChannelSelector());
-                }
-                return null;
-            })
-            .collect(Collectors.toList());
-    }
-
-    @Named("fromSelectors")
-    default List<Selector> fromSelectors(List<io.gravitee.definition.model.v4.flow.selector.Selector> selectors) {
-        if (Objects.isNull(selectors)) {
-            return new ArrayList<>();
-        }
-        return selectors
-            .stream()
-            .map(selector -> {
-                if (selector == null) {
-                    return null;
-                }
-                if (selector.getType() == SelectorType.HTTP) {
-                    return new io.gravitee.rest.api.management.v2.rest.model.Selector(
-                        this.map((io.gravitee.definition.model.v4.flow.selector.HttpSelector) selector)
-                    );
-                }
-                if (selector.getType() == SelectorType.CONDITION) {
-                    return new io.gravitee.rest.api.management.v2.rest.model.Selector(
-                        this.map((io.gravitee.definition.model.v4.flow.selector.ConditionSelector) selector)
-                    );
-                }
-                if (selector.getType() == SelectorType.CHANNEL) {
-                    return new io.gravitee.rest.api.management.v2.rest.model.Selector(
-                        this.map((io.gravitee.definition.model.v4.flow.selector.ChannelSelector) selector)
-                    );
-                }
-                return null;
-            })
-            .collect(Collectors.toList());
-    }
-
-    // DateTime
-    default OffsetDateTime map(Date value) {
-        if (Objects.isNull(value)) {
-            return null;
-        }
-        return value.toInstant().atOffset(ZoneOffset.UTC);
     }
 }
